updated fetching added detailed summary
set up the fetching to better suit the needs of the app. I was having a lot of issues keeping things modular but now I've got everything handled fetching wise in Main, which is helpful on that front. Also added a summary panel for when user clicks on a pokemon. Added id numbers and names to cards. In the process of refactoring, search no longer functions so TODO: fix search.